get all images in DB with the new structure of person_bb

+++ dbUtils.py
@@ -932,7 +932,7 @@ def reconstruct_db_images(images_collection):
         for person in doc['people']:
             if len(person['face']):
                 x, y, w, h = person['face']
-                person_bb = [round(max(0, x - 1.5 * w)), y, round(min(image.shape[1], x + 2.5 * w)),
+                person_bb = [int(round(max(0, x - 1.5 * w))), y, int(round(min(image.shape[1], x + 2.5 * w))),
                              min(image.shape[0], 8 * h)]
             else:
                 person_bb = None
